+++ lib/GRNOC/NetSage/Deidentifier/FlowMover.pm
@@ -1,4 +1,4 @@
-package GRNOC::NetSage::Anonymizer::FlowMover;
+package GRNOC::NetSage::Deidentifier::FlowMover;
 
 use strict;
 use warnings;
@@ -6,7 +6,7 @@ use warnings;
 
 use Moo;
 
-extends 'GRNOC::NetSage::Anonymizer::Pipeline';
+extends 'GRNOC::NetSage::Deidentifier::Pipeline';
 
 use GRNOC::Log;
 use GRNOC::Config;
@@ -40,7 +40,7 @@ sub BUILD {
 # expects an array of data for it to process
 # in this case we want to copy the messages unmodified, so returns the original data
 sub _process_messages {
-    # TODO: the actual anonymization in a better way
+    # TODO: the actual deidentification in a better way
     my ( $self, $messages ) = @_;
 
     #my $finished_messages = $messages;