+++ CHANGES.md
@@ -1,5 +1,5 @@
-## GRNOC NetSage Anonymizer 0.0.1 -- Fri April 1 2016
+## GRNOC NetSage Deidentifier 0.0.1 -- Fri April 1 2016
 
 ### Features:
- * ISSUE=543 PROJ=160 Initial implementation of NetSage data aggregation process. 
+ * ISSUE=543 PROJ=160 Initial implementation of NetSage flow deidentification pipeline.